Fix creating project
