+++ web/src/components/atoms/TextArea/index.tsx
@@ -1,0 +1,3 @@
+import TextArea from "antd/lib/input/TextArea";
+
+export default TextArea;
