+++ web/src/components/molecules/Dashboard/ProjectList.tsx
@@ -14,14 +14,14 @@ export interface Props {
   className?: string;
   projects?: ProjectType[];
   workspaceId?: string;
-  handleProjectModalOpen: () => void;
+  // handleProjectModalOpen: () => void;
 }
 
 const ProjectList: React.FC<Props> = ({
   className,
   projects,
   workspaceId,
-  handleProjectModalOpen,
+  // handleProjectModalOpen,
 }) => {
   const navigate = useNavigate();
   const { Meta } = Card;
@@ -30,7 +30,6 @@ const ProjectList: React.FC<Props> = ({
       {projects?.length ? (
         <Content>
           {projects.map((project) => (
-            // <Project key={project.id} project={project} />
             <CardWrapper key={project.id}>
               <Card
                 cover={<Cover>{project.name.charAt(0)}</Cover>}
@@ -56,7 +55,7 @@ const ProjectList: React.FC<Props> = ({
           <Suggestion>
             Create a new project{" "}
             <Button
-              onClick={handleProjectModalOpen}
+              // onClick={handleProjectModalOpen}
               type="primary"
               icon={<PlusOutlined />}
             >
