+++ web/src/components/organisms/Project/hooks.ts
@@ -1,20 +1,12 @@
 import { Project } from "@reearth-cms/components/molecules/Dashboard/types";
 import {
   useGetProjectsQuery,
-  useUpdateProjectMutation,
-  useDeleteProjectMutation,
   useCreateProjectMutation,
 } from "@reearth-cms/gql/graphql-client-api";
 import { useWorkspace } from "@reearth-cms/state";
 import { useCallback, useMemo, useState } from "react";
-import { useNavigate } from "react-router-dom";
 
-type Params = {
-  projectId?: string;
-};
-
-export default ({ projectId }: Params) => {
-  const navigate = useNavigate();
+export default () => {
   const [currentWorkspace] = useWorkspace();
   const [projectModalShown, setProjectModalShown] = useState(false);
 
@@ -64,66 +56,6 @@ export default ({ projectId }: Params) => {
     [createNewProject, workspaceId, refetch]
   );
 
-  const rawProject = useMemo(
-    () => data?.projects.nodes.find((p: any) => p?.id === projectId),
-    [data, projectId]
-  );
-  const project = useMemo(
-    () =>
-      rawProject?.id
-        ? {
-            id: rawProject.id,
-            name: rawProject.name,
-            description: rawProject.description,
-            alias: rawProject.alias,
-          }
-        : undefined,
-    [rawProject]
-  );
-
-  const [updateProjectMutation] = useUpdateProjectMutation();
-  const [deleteProjectMutation] = useDeleteProjectMutation({
-    refetchQueries: ["GetProjects"],
-  });
-
-  const handleProjectUpdate = useCallback(
-    (data: { name?: string; description: string }) => {
-      if (!projectId || !data.name) return;
-      updateProjectMutation({
-        variables: {
-          projectId,
-          name: data.name,
-          description: data.description,
-        },
-      });
-    },
-    [projectId, updateProjectMutation]
-  );
-
-  const handleProjectDelete = useCallback(async () => {
-    if (!projectId) return;
-    const results = await deleteProjectMutation({ variables: { projectId } });
-    if (results.errors) {
-      console.log("errors");
-    } else {
-      console.log("succeed");
-      navigate(`/dashboard/${workspaceId}`);
-    }
-  }, [projectId, deleteProjectMutation, navigate, workspaceId]);
-
-  const [assetModalOpened, setOpenAssets] = useState(false);
-
-  const toggleAssetModal = useCallback(
-    (open?: boolean) => {
-      if (!open) {
-        setOpenAssets(!assetModalOpened);
-      } else {
-        setOpenAssets(open);
-      }
-    },
-    [assetModalOpened, setOpenAssets]
-  );
-
   const handleProjectModalClose = useCallback(
     (r?: boolean) => {
       setProjectModalShown(false);
@@ -140,16 +72,11 @@ export default ({ projectId }: Params) => {
   );
 
   return {
-    project,
     projects,
-    projectId,
     currentWorkspace,
+    projectModalShown,
     handleProjectCreate,
-    handleProjectUpdate,
-    handleProjectDelete,
     handleProjectModalOpen,
     handleProjectModalClose,
-    assetModalOpened,
-    toggleAssetModal,
   };
 };
